requirements.txt
setproctitle
opencv-python
numpy
RPI.GPIO
pyserial
tflite-runtime
Pillow
Flask